requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
websockets==12.0
google-generativeai==0.3.2
opencv-python==4.9.0.80
mediapipe==0.10.9
python-dotenv==1.0.1